break up init helpers
